tests/wpt/tests/xhr/formdata/constructor-submitter-coordinate.html

<!DOCTYPE html>
<meta charset='utf-8'>
<meta name=viewport content='width=device-width,initial-scale=1'>
<link rel='help' href='https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#constructing-the-form-data-set'>
<link ref='help' href='https://xhr.spec.whatwg.org/#dom-formdata'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>

<form>
  <input type=image src='/images/100px-green-rect.svg'>
</form>

<script>
"use strict";

promise_test(async () => {
  await new Promise(r => window.addEventListener("load", r));

  const form = document.querySelector("form");
  const submitter = form.querySelector("input[type=image]");

  let formData;
  form.addEventListener("submit", e => {
    e.preventDefault();
    formData = new FormData(form, submitter);
  });

  const domRect = submitter.getBoundingClientRect();
  const clientX = Math.round((domRect.left + domRect.right) / 7);
  const clientY = Math.round((domRect.top + domRect.bottom) / 4);

  submitter.dispatchEvent(
    new MouseEvent("click", {clientX, clientY})
  );

  assert_equals(formData?.get("x"), "23");
  assert_equals(formData?.get("y"), "46");
}, "The constructed FormData object should contain correct entries for Image Button submitter's dispatched coordinate");
</script>
